requirements.txt
keras==3.10.0
matplotlib==3.10.3
numpy==2.3.0
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
pandas==2.3.0
Pillow==11.2.1
streamlit==1.38.0
tensorflow==2.19.0
tensorflow_intel==2.11.0
